refactor(cluster): propagate VShardId u32 through control plane and server
Complete the u16→u32 widening for vshard identifiers in the control
plane: pgwire session state, ILP listener shard dispatch, CDC stream
handlers, gateway dispatch, WAL replication encode path, cluster
metrics, and test helpers.

Also derive `ToMessagePack`/`FromMessagePack` on `PartitionMeta` and
`PartitionState` in `nodedb-types` so timeseries partitions can be
serialized over the cluster wire protocol without a manual conversion.

+++ nodedb/src/control/catalog_entry/post_apply/async_dispatch/materialized_view.rs
@@ -32,7 +32,7 @@ pub async fn delete_async(tenant_id: u32, name: String, shared: Arc<SharedState>
             let request = Request {
                 request_id,
                 tenant_id: TenantId::new(tenant_id),
-                vshard_id: VShardId::new(core_id as u16),
+                vshard_id: VShardId::new(core_id as u32),
                 plan: PhysicalPlan::Meta(MetaOp::UnregisterMaterializedView {
                     tenant_id,
                     name: name.clone(),
